requirements.txt
pymysql
ConfigParser
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
sqlalchemy_utils
passlib
Flask-Testing==0.6.2
coverage==4.5.1
gunicorn==19.8.1
flask-migrate==2.2.0
Flask-Testing==0.6.2
coverage==4.5.1